removed int cast

+++ mean-reversion/arima_with_mean_reversion_v1.py
@@ -248,7 +248,7 @@ class Trader:
                 for Trade in state.market_trades.get(product, []):
                     price += Trade.price * Trade.quantity
                     count += Trade.quantity
-                current_avg_market_price = int(price / count)
+                current_avg_market_price = price / count
                 price_history_pearl = np.append(price_history_pearl, current_avg_market_price)
                 if len(price_history_pearl) == history_length+1:
                     price_history_pearl = price_history_pearl[1:]
@@ -290,7 +290,7 @@ class Trader:
                 for Trade in state.market_trades.get(product, []):
                     price += Trade.price * Trade.quantity
                     count += Trade.quantity
-                current_avg_market_price = int(price / count)
+                current_avg_market_price = price / count
                 price_history_banana = np.append(price_history_banana, current_avg_market_price)
                 if len(price_history_banana) == history_length+1:
                     price_history_banana = price_history_banana[1:]
